♻️ (topology): Extract bucketForBudget to dedicated module for react-refresh
